refactor(BridgeTest) : 패키지 이동
BridgeTest 패키지 이동
TestBridgeGenerator 클래스 생성
